refactored graphics to different classes
