+++ ZombustersWindows/MainScreens/BackgroundScreen.cs
@@ -1,99 +1,12 @@
-﻿#region File Description
-//-----------------------------------------------------------------------------
-// Game1.cs
-//
-// Microsoft XNA Community Game Platform
-// Copyright (C) Microsoft Corporation. All rights reserved.
-//-----------------------------------------------------------------------------
-#endregion
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
 using GameStateManagement;
 using Microsoft.Xna.Framework.Input.Touch;
 
 namespace ZombustersWindows
 {
-    public sealed class Explosion
-    {
-        private static Texture2D texture;
-        public static Texture2D Texture
-        {
-            get { return texture; }
-            set { texture = value; }
-        }
-        private Explosion() { }
-        public static void Draw(SpriteBatch batch, Vector2 pos, float rotation, double totalSeconds, double startSeconds, double endSeconds)
-        {
-            // If the explosion is finished, or hasn't started yet, return
-            if ((totalSeconds > endSeconds) || (startSeconds > totalSeconds))
-                return;
-
-            double duration = endSeconds - startSeconds;
-            double time = totalSeconds - startSeconds;
-            
-            // The explosion gets bigger as time goes on
-            float scale = (float) ((time / duration)*.5f) + .6f;
-
-            Vector2 origin = new Vector2(texture.Height / 2);
-            
-            // The explosion gets brighter as time goes on
-            float intensity = (float)(time / duration) * .5f;
-            Vector4 colorval = new Vector4(0.75f + intensity);
-
-            // The explosion also fades in and out
-            colorval.W = (float)Math.Sin((time / duration) * MathHelper.Pi)+.2f;
-
-            // The intensity and fade determines our tint color
-            Color color = new Color(colorval);
-
-            // Draw the texture with the appropriate scale and tint
-            batch.Draw(texture, pos, null, color, rotation, origin, scale, SpriteEffects.None, 0.5f);            
-        }
-    }
-
-    public class ScrollingBackground
-    {
-        private Vector2 screenpos, origin, texturesize;
-        private Texture2D mytexture;
-        private int screenheight;
-        public void Load(GraphicsDevice device, Texture2D backgroundTexture)
-        {
-            mytexture = backgroundTexture;
-            screenheight = device.Viewport.Height;
-            int screenwidth = device.Viewport.Width;
-            // Set the origin so that we're drawing from the 
-            // center of the top edge.
-            origin = new Vector2(mytexture.Width / 2, 0);
-            // Set the screen position to the center of the screen.
-            screenpos = new Vector2(screenwidth / 2, screenheight / 2);
-            // Offset to draw the second texture, when necessary.
-            texturesize = new Vector2(0, mytexture.Height);
-        }
-        public void Update(float deltaY)
-        {
-            screenpos.Y += deltaY;
-            screenpos.Y = screenpos.Y % mytexture.Height;
-        }
-        public void Draw(SpriteBatch batch, Color color)
-        {
-            // Draw the texture, if it is still onscreen.
-            if (screenpos.Y < screenheight)
-            {
-                batch.Draw(mytexture, screenpos, null,
-                     color, 0, origin, 1, SpriteEffects.None, 0f);
-            }
-            // Draw the texture a second time, behind the first,
-            // to create the scrolling illusion.
-            batch.Draw(mytexture, screenpos - texturesize, null,
-                 color, 0, origin, 1, SpriteEffects.None, 0f);
-        }
-    }
-
     public class BackgroundScreen : GameScreen
     {
         ScrollingBackground stars;
