+++ ZombustersWindows/ZombustersWindows.csproj
@@ -67,6 +67,7 @@
     <Compile Include="BloomPostProcess\BloomSettings.cs" />
     <Compile Include="ContentManager\LevelManager.cs" />
     <Compile Include="Enemies.cs" />
+    <Compile Include="ContentManager\Explosion.cs" />
     <Compile Include="GameStateManagement\GameState.cs" />
     <Compile Include="MainScreens\LoadInScreen.cs" />
     <Compile Include="GameStateManagement\MessageType.cs" />
@@ -77,7 +78,7 @@
     <Compile Include="GameStateManagement\IScreenFactory.cs" />
     <Compile Include="GameStateManagement\PlayerIndexEventArgs.cs" />
     <Compile Include="GameStateManagement\ScreenManager.cs" />
-    <Compile Include="Graphics.cs" />
+    <Compile Include="MainScreens\BackgroundScreen.cs" />
     <Compile Include="MainScreens\GameOverMenu.cs" />
     <Compile Include="MainScreens\GamePlayMenu.cs" />
     <Compile Include="MainScreens\GamePlayScreen.cs" />
@@ -107,6 +108,7 @@
     <Compile Include="Player.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="MainScreens\ScrollingBackground.cs" />
     <Compile Include="SteeringBehaviors\Arrive.cs" />
     <Compile Include="SteeringBehaviors\Circle.cs" />
     <Compile Include="SteeringBehaviors\GameWorld.cs" />
